+++ AuthorizerPresentation/Controllers/UserController.cs
@@ -1,5 +1,6 @@
 ﻿using AuthorizerBLL.Services;
 using AuthorizerPresentation.ViewModels;
+using Common.DataTransferObjects;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -10,6 +11,7 @@ using System.Web.Mvc;
 
 namespace AuthorizerPresentation.Controllers
 {
+    [Authorize(Roles = "admin")]
     public class UserController : Controller
     {
         private readonly IUserService _userService;
@@ -20,31 +22,71 @@ namespace AuthorizerPresentation.Controllers
             _userService = userService;
             _roleService = roleService;
         }
+
+        /// <summary>
+        /// Display list of users
+        /// </summary>
+        /// <returns></returns>
         public ActionResult Index()
         {
-            IList<UserViewModel> users = _userService.FindAll().Select(b => new UserViewModel(){
-                FirstName = b.FirstName,
-                LastName = b.LastName,
-                Password = b.Password,
-                UserName = b.UserName,
-                UserRole = b.Role
-            }).ToList();
-            return View(users);
+            var userDtoList = _userService.FindAll();
+
+            if (userDtoList != null)
+            {
+                //Mapping UserDto to UserViewModel
+                IList<UserViewModel> users = userDtoList.Select(b => new UserViewModel()
+                {
+                    FirstName = b.FirstName,
+                    LastName = b.LastName,
+                    Password = b.Password,
+                    UserName = b.UserName,
+                    RoleId = b.RoleId,
+                    RoleName = FetchRoles(b.RoleId)
+                }).ToList();
+                return View(users);
+            }
+            else
+            {
+                return View();
+            }
+            
         }
+
+        //Create a new user
         public ActionResult Create()
         {
-            FetchRoles();
+            ViewData["Roles"] = FetchRoles();
             return View();
         }
         [HttpPost]
         public ActionResult Create(UserViewModel userView)
         {
-            var a = userView.UserRole.RoleName;        
+            if (userView != null)
+            {
+                //Refreshes roles to DropDown on the case that duplicate username error occurs
+                ViewData["Roles"] = FetchRoles();
 
-            _userService.Create(userView);
-            return RedirectToAction("Index");
+                userView.FirstName.Trim();
+                var userCreateResult = _userService.Create(userView);
+
+                if (userCreateResult == -1)
+                {
+                    ModelState.AddModelError("UserName", "User Name Already Exists");
+                    return View(userView);
+                }
+                else if (userCreateResult == 0)
+                {
+                    ViewBag.Message = "Db Creation Error! Please Restart Application";
+                }
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
         }
 
+        //Edit existing user
         public ActionResult Edit(string username)
         {
             if (username == null)
@@ -56,7 +98,7 @@ namespace AuthorizerPresentation.Controllers
             {
                 return HttpNotFound();
             }
-            FetchRoles();
+            ViewData["Roles"] = FetchRoles();
             return View(userCollection);
         }
 
@@ -68,31 +110,65 @@ namespace AuthorizerPresentation.Controllers
             return RedirectToAction("Index");
         }
 
+        //Delete a user
         public ActionResult Delete(string username)
         {
-            UserViewModel userCollection = _userService.GetByUserName(username);
-            return View(userCollection);
+            if (username == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            else
+            {
+                UserViewModel userCollection = _userService.GetByUserName(username);
+                if (HttpContext.User.Identity.Name == username)
+                {
+                    ViewBag.Error = "Cannot delete yourself";
+                    return View();
+                }
+                else
+                {
+                    return View(userCollection);
+                }
+            }
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Delete(UserViewModel userDetails, FormCollection requestValidator)
         {
-            _userService.Delete(userDetails);
-            return RedirectToAction("Index");
+            if (userDetails == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            else
+            {
+                _userService.Delete(userDetails);
+                return RedirectToAction("Index");
+            }
         }
 
         /// <summary>
         /// Fetches list of all roles and saves to ViewData["Roles"]
         /// </summary>
-        private void FetchRoles()
+        private SelectList FetchRoles()
         {
             ViewData.Clear();
 
             //Obtaining list of roles
             var roleList = _roleService.FindAll().ToList();
 
-            //Saving to viewdata for populating combo box
-            ViewData["Roles"] = new SelectList(roleList.Select(r => r.RoleName), "roleName");
+            //Setting to select list for populating combo box
+            return new SelectList(roleList, "roleId", "roleName");
+        }
+
+        //Fetches role based on id
+        private string FetchRoles(int id)
+        {
+            ViewData.Clear();
+
+            //Obtaining list of roles
+            var roleList = _roleService.FindAll().ToList();
+
+            return roleList.Where(r => r.RoleId == id).Select(r => r.RoleName).First();
         }
     }
 }
